requirements.txt
flask==3.0.0
flask-cors==4.0.0
torch==2.1.0
torchvision==0.16.0
diffusers==0.25.0
transformers==4.36.0
accelerate==0.25.0
safetensors==0.4.1
Pillow==10.1.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
scipy==1.11.4
gunicorn==21.2.0
